moved the root route to own component to remove routing warnings w/ burnettk
